Extract and re-enable deprecated props tests
