+++ .vscode/settings.json
@@ -8,7 +8,7 @@
     "*.mjs": "${capture}.mjs.map, ${capture}.min.mjs, ${capture}.d.ts",
     "*.css": "${capture}.css.map, ${capture}.min.css",
     "*.jsx": "${capture}.js",
-    "*.tsx": "${capture}.ts, ${capture}.test.tsx",
+    "*.tsx": "${capture}.js, ${capture}.ts, ${capture}.d.ts, ${capture}.test.ts, ${capture}.test.tsx, ${capture}.*.test.ts, ${capture}.*.test.tsx",
     "tsconfig.json": "tsconfig.*.json",
     "package.json": "package-lock.json, .npmrc, yarn.lock, .yarnrc",
     "vite.config.js": "vite.config.*.js",
