+++ packages/react-querybuilder/src/RuleGroup.test.tsx
@@ -7,15 +7,9 @@ import type {
 import { render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import { consoleMocks, createRule, getRuleGroupProps, ruleGroupClassnames } from '../genericTests';
-import {
-  defaultCombinators,
-  defaultTranslations as t,
-  standardClassnames as sc,
-  TestID,
-} from './defaults';
-import { errorDeprecatedRuleGroupProps, errorEnabledDndWithoutReactDnD } from './messages';
+import { defaultTranslations as t, standardClassnames as sc, TestID } from './defaults';
+import { errorEnabledDndWithoutReactDnD } from './messages';
 import { RuleGroup } from './RuleGroup';
-import { add } from './utils';
 
 const user = userEvent.setup();
 
@@ -429,45 +423,6 @@ describe('dynamic classNames', () => {
       'custom-groupBased-class',
       'custom-combinatorBased-class'
     );
-  });
-});
-
-describe('deprecated props', () => {
-  // TODO: We're invoking the useDeprecatedProps module twice (see two `it` calls below),
-  // so we may need to use https://www.npmjs.com/package/babel-plugin-dynamic-import-node
-  // to reset the module import (to reset the "has already warned about this" switch
-  // in the useDeprecatedProps module). Skipping this one until we can resolve the issue.
-  it.skip('warns about deprecated props', () => {
-    // @ts-expect-error ruleGroup is required
-    render(<RuleGroup {...getRuleGroupProps()} ruleGroup={undefined} rules={[]} combinator="or" />);
-    expect(consoleError).toHaveBeenCalledWith(errorDeprecatedRuleGroupProps);
-    expect(screen.getByTestId(TestID.combinators)).toHaveValue('or');
-  });
-
-  it('warns about deprecated props (independent combinators)', async () => {
-    const addListener = jest.fn();
-    render(
-      <RuleGroup
-        {...getRuleGroupProps(
-          { independentCombinators: true },
-          {
-            onRuleAdd: (rOrG, parentPath) => {
-              addListener(
-                add({ rules: [{ field: 'f', operator: '=', value: 'v' }] }, rOrG, parentPath)
-              );
-            },
-          }
-        )}
-        path={[]}
-        // @ts-expect-error ruleGroup prop is required
-        ruleGroup={undefined}
-        rules={[{ field: 'f', operator: '=', value: 'v' }]}
-        combinator={undefined}
-      />
-    );
-    expect(consoleError).toHaveBeenCalledWith(errorDeprecatedRuleGroupProps);
-    await user.click(screen.getByTestId(TestID.addRule));
-    expect(addListener.mock.calls[0][0].rules[1]).toBe(defaultCombinators[0].name);
   });
 });
